feat: add support for recording intermediate images during CT to mesh conversion
This commit adds support for recording intermediate images during the CT to mesh conversion process. The `main` function now includes a `counter` variable that keeps track of the number of images saved. The `save_img` function is called at various points in the code to save the current data as an image. The images are saved in a directory specified by the `record_dir` argument.

The `main` function now also includes a `component` argument, which is an instance of the `Component` enum. This argument is used to determine the threshold value for the data based on the selected component. The threshold values are defined in the `THRESHOLDS` dictionary.

The code has been updated to handle the different components separately. The `match` statement is used to select the appropriate code block based on the selected component. The code for each component includes the steps to remove the background, find the largest object, and save intermediate images.

The `main` function now takes an optional `record_dir` argument, which specifies the directory where the intermediate images should be saved. If this argument is not provided, the images will not be saved.

The `main` function now also includes logging statements to output the number of features found during the processing of each component.

The `main` function has been updated to export the resulting mesh using the `export_voxel` function.

The `main` function now takes an additional `component` argument, which is used to determine the threshold value for the data based on the selected component. The threshold values are defined in the `THRESHOLDS` dictionary.

The `main` function now includes a `counter` variable that keeps track of the number of images saved. The `save_img` function is called at various points in the code to save the current data as an image. The images are saved in a directory specified by the `record_dir` argument.

The code has been updated to handle the different components separately. The `match` statement is used to select the appropriate code block based on the selected component. The code for each component includes the steps to remove the background, find the largest object, and save intermediate images.

The `export_voxel` function has been updated to export the resulting mesh.

The `main` function now includes logging statements to output the number of features found during the processing of each component.

The `main` function now takes an optional `