feat: update correspondence scale and weight normal in registration
Update the correspondence scale and correspondence weight normal in the registration process to improve the accuracy and stability of the mesh registration. By reducing the correspondence scale from 1.05 to 0.01 and the correspondence weight normal from 0.5 to 0.05, the registration process will be more precise and robust, leading to better alignment of the meshes. Additionally, the calculation of input points has been adjusted to incorporate the correspondence scale and input normals, further enhancing the registration process.